Web_Application/requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
pymongo
flask-login
